refactored all controller files into routes files
